Restructure + fix bugs
