+++ bot/config.txt
@@ -1,0 +1,1 @@
+{'model': '{0: <pybamm.models.full_battery_models.lithium_ion.spme.SPMe object at 0x7fdf633bbe20>, 1: <pybamm.models.full_battery_models.lithium_ion.dfn.DFN object at 0x7fdf62e2c220>, 2: <pybamm.models.full_battery_models.lithium_ion.spm.SPM object at 0x7fdf62ac3580>}', 'model options': None, 'chemistry': {'chemistry': 'lithium_ion', 'cell': 'kokam_Marquis2019', 'negative electrode': 'graphite_mcmb2528_Marquis2019', 'separator': 'separator_Marquis2019', 'positive electrode': 'lico2_Marquis2019', 'electrolyte': 'lipf6_Marquis2019', 'experiment': '1C_discharge_from_full_Marquis2019', 'sei': 'example', 'citation': 'Marquis2019'}, 'is_experiment': False, 'cycle': None, 'number': None, 'is_comparison': True, 'param_to_vary': None, 'varied_values': []}
