+++ bot/data.txt
@@ -17,3 +17,5 @@
 2021-07-23 19:09:54 {'model': '{0: <pybamm.models.full_battery_models.lithium_ion.spme.SPMe object at 0x7f5c75cb0e80>}', 'model options': None, 'chemistry': {'chemistry': 'lithium_ion', 'cell': 'kokam_Marquis2019', 'negative electrode': 'graphite_mcmb2528_Marquis2019', 'separator': 'separator_Marquis2019', 'positive electrode': 'lico2_Marquis2019', 'electrolyte': 'lipf6_Marquis2019', 'experiment': '1C_discharge_from_full_Marquis2019', 'sei': 'example', 'citation': 'Marquis2019'}, 'is_experiment': False, 'cycle': None, 'number': None, 'is_comparison': True, 'param_to_vary': 'Ambient temperature [K]', 'varied_values': [311.02, 333.07]}
 2021-07-24 07:20:40 {'model': '{0: <pybamm.models.full_battery_models.lithium_ion.dfn.DFN object at 0x7f9296c90640>, 1: <pybamm.models.full_battery_models.lithium_ion.spm.SPM object at 0x7f929673cf70>}', 'model options': None, 'chemistry': {'chemistry': 'lithium_ion', 'cell': 'Yang2017', 'negative electrode': 'graphite_Yang2017', 'separator': 'separator_Yang2017', 'positive electrode': 'nmc_Yang2017', 'electrolyte': 'lipf6_Ecker2015', 'experiment': '1C_discharge_from_full_Ecker2015', 'sei': 'yang2017_sei', 'lithium plating': 'yang2017_Li_plating', 'citation': ['Yang2017', 'Ecker2015i']}, 'is_experiment': True, 'cycle': [('Discharge at 3 C until 3.5 V', 'Rest for 3 minutes', 'Charge at 3 C until 3.9 V', 'Hold at 3.9 V until 50 mA')], 'number': 2, 'is_comparison': True, 'param_to_vary': None, 'varied_values': []}
 2021-07-24 19:59:41 {'model': '{0: <pybamm.models.full_battery_models.lithium_ion.spm.SPM object at 0x7f87c0b242b0>, 1: <pybamm.models.full_battery_models.lithium_ion.dfn.DFN object at 0x7f87c05def40>}', 'model options': None, 'chemistry': {'chemistry': 'lithium_ion', 'cell': 'Enertech_Ai2020', 'negative electrode': 'graphite_Ai2020', 'separator': 'separator_Ai2020', 'positive electrode': 'lico2_Ai2020', 'electrolyte': 'lipf6_Enertech_Ai2020', 'experiment': '1C_discharge_from_full_Ai2020', 'sei': 'example', 'citation': 'Ai2019'}, 'is_experiment': True, 'cycle': [('Discharge at 3 C until 3.6 V', 'Charge at 2 C until 4.1 V', 'Hold at 4.1 V until 4 mA')], 'number': 1, 'is_comparison': True, 'param_to_vary': None, 'varied_values': []}
+2021-07-25 07:23:45 {'model': '{0: <pybamm.models.full_battery_models.lithium_ion.dfn.DFN object at 0x7f1d6edb4070>, 1: <pybamm.models.full_battery_models.lithium_ion.spme.SPMe object at 0x7f1d6e845dc0>}', 'model options': None, 'chemistry': {'chemistry': 'lithium_ion', 'cell': 'LGM50_Chen2020', 'negative electrode': 'graphite_Chen2020', 'separator': 'separator_Chen2020', 'positive electrode': 'nmc_Chen2020', 'electrolyte': 'lipf6_Nyman2008', 'experiment': '1C_discharge_from_full_Chen2020', 'sei': 'example', 'citation': 'Chen2020'}, 'is_experiment': False, 'cycle': None, 'number': None, 'is_comparison': True, 'param_to_vary': None, 'varied_values': []}
+2021-07-25 19:31:21 {'model': '{0: <pybamm.models.full_battery_models.lithium_ion.spme.SPMe object at 0x7fdf633bbe20>, 1: <pybamm.models.full_battery_models.lithium_ion.dfn.DFN object at 0x7fdf62e2c220>, 2: <pybamm.models.full_battery_models.lithium_ion.spm.SPM object at 0x7fdf62ac3580>}', 'model options': None, 'chemistry': {'chemistry': 'lithium_ion', 'cell': 'kokam_Marquis2019', 'negative electrode': 'graphite_mcmb2528_Marquis2019', 'separator': 'separator_Marquis2019', 'positive electrode': 'lico2_Marquis2019', 'electrolyte': 'lipf6_Marquis2019', 'experiment': '1C_discharge_from_full_Marquis2019', 'sei': 'example', 'citation': 'Marquis2019'}, 'is_experiment': False, 'cycle': None, 'number': None, 'is_comparison': True, 'param_to_vary': None, 'varied_values': []}
