--- bot/twitter_api/config.txt
@@ -1,1 +1,0 @@
-{'model': '{0: <pybamm.models.full_battery_models.lithium_ion.spm.SPM object at 0x7f87c0b242b0>, 1: <pybamm.models.full_battery_models.lithium_ion.dfn.DFN object at 0x7f87c05def40>}', 'model options': None, 'chemistry': {'chemistry': 'lithium_ion', 'cell': 'Enertech_Ai2020', 'negative electrode': 'graphite_Ai2020', 'separator': 'separator_Ai2020', 'positive electrode': 'lico2_Ai2020', 'electrolyte': 'lipf6_Enertech_Ai2020', 'experiment': '1C_discharge_from_full_Ai2020', 'sei': 'example', 'citation': 'Ai2019'}, 'is_experiment': True, 'cycle': [('Discharge at 3 C until 3.6 V', 'Charge at 2 C until 4.1 V', 'Hold at 4.1 V until 4 mA')], 'number': 1, 'is_comparison': True, 'param_to_vary': None, 'varied_values': []}
